Implemented start game
Within the multiplayer with user can now click the start game button. This will lead to a request with the action startGame. The backend handles it.

Missing is the correct response handling in the client due to missing implementation in the backend. In the end every client should see a new Gameboard view if any client started the game.