Adapted start game
Now every client can press the startGame button. This will notify the backend to start the game if not already. The clicking client will be informed if the setup was successfully. All Clients will get a message gameIsStarted. Added handler to react to this message by switching to the placeholder game view.